Got tutorial 7 in a working state
Currently skipped as always.

Also removed some unused pieces of code:
- Imports
- Outer test suite that didn't do anything useful
